Refactor the way configuration works.
This commit does change the way configuration is done.
We no longer ship with a default ``config.ini``. Instead we provide a
new ``_write_config_file`` function that will write a vanilla
``config.ini`` to a location passed.
If the location dir does not exist it will create it.

Upon ``hamster-cli`` invocation, we now call a new
``_get_config_instance`` function that is tasked with any filesystem
related logic and is to provide the actual SafeConfigParser instance.
For that purpose it checks its default locations for an existing
``hamster_cli.conf`` file and calls the new ``_write_config_file`` if it can
not find one. Either way, it will return a ``SafeConfigParser``
instance.

With this basis, our existing ``_get_config`` function now focuses on
turning the ``SafeConfigParser`` instance into a suitable and complete
config dir apropiate for consumption by our client and backend.