Refactored ``_get_config``.
Doe to our changes to how configuration works, adjustments to
``_get_config`` were in order. I used the oportunity to turn each
value-retrieval into a single function. This way additional checks and
logic will be confined to those inner funktions and our main code stays
clean and compartmentalized.
This approach has the added benifit that it allows us do deal with the
temporary need to still deal with ``work_dir`` keys in a bearable way
that is not spoiling our entire codebase.